[Refact]#19 - 메인페이지 훅 분리
